requirements.txt
allure-pytest~=2.9.45
allure-python-commons~=2.9.45
mimesis~=5.2.1
pytest~=6.2.5
requests~=2.26.0
selenium~=4.1.0